requirements.txt
Flask
pymongo
python-dotenv
authlib
requests
Flask-Mail
pyotp
qrcode
pycryptodome
numpy
Pillow